server/requirements.txt
aniso8601==8.0.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
httpie==2.4.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
PySocks==1.7.1
pytz==2020.1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.26.4
Werkzeug==1.0.1
